Disseretstion
In modern era of conducting any business, customer satisfaction has become a must. There is a high competition in every industry either in retail businesses or in e-commerce. No matter what products or services are being offered by a firm there is a high chance that some other firm offering the same. If a company wants to sustain for a longer period it must offer the products or services at right price. That right price will make the customers happy in addition by offering product or services at right pricing the company will achieve a certain competitive advantage over other companies operating in the same industry. This study aims to build a price prediction model utilizing machine learning algorithm on Dynamic Pricing Strategy to have the right price. The model shall be able to predict price in real-time depending on multiple factors responsible for the change in price.